Fix incorrect veneer generation by GNU ld
The linker produced a branch to a veneer function using the T2 variant
of the BLX instruction, which isn't supported by Armv8-M. According to
[1], the linker exhibits a weird behavior when "you are branching to
something which is not marked as being a function".

It turned out to be correct. This commit fixes the issue by marking the
import symbols as functions.

[1]: https://bugs.launchpad.net/gcc-arm-embedded/+bug/1697255