example: Add a new example `bench-latency`
This is a Zig port of the benchmark program that exists in
b042babe64980480a511035d8940a19513849aee (in `master`).

I made a slight modification for it to run more efficiently.
Specifically, the delay loop for waiting for both timers to fire was
replaced with a loop that polls the completion flag. Though I'm afraid
that this might interfere with the measurement...