Add `__TCPrivateShadowAssertReturnFast`
In C, it's quite rare to use such a return type that use these
registers, so the register spill/restore overhead in
`__TCPrivateShadowAssertReturn` can be eliminated by killing these
registers when possible.

A new private gateway function `__TCPrivateShadowAssertReturnFast` is
defined. When these registers are not used for a return value, this new
gateway function is used instead.