refactor(examples): move TZmCFI Monitor to a separate compilation unit
I'm planning to move performance-critical code to RAM for faster
execution. To do this, the linker must be able to distinguish this
section of code from the rest. So, in this commit, I moved TZmCFI
Monitor to a separate compilation unit, which is to be linked to the
main executable as a static library. The linker script can now specify
this code by the name `monitor.o`.