fix(examples): pass `-mcpu=cortex-m33` to clang when compiling `*.S`
The linker guesses the target CPU and features by examining the input
files' ARM attributes, so it's important that we set appropriate
attributes when compiling object files.

Zig invokes Clang to compile C/C++ and assembler source files. In order
to specify a target CPU and features, it passes `-Xclang -target-cpu
...` to Clang. Unfortunately, it doesn't do it for assembler source
files, and there's no way to get this working in Zig.

This commit works around this problem by passing `-mcpu=cortex-m33` to
Clang, which will transform it into lower-level options `-target-cpu`
and `-target-feature`.